Moved Decoy Generation
