+++ mzLib/Test/DatabaseTests/TestProteomicsReadWrite.cs
@@ -12,6 +12,7 @@ using Proteomics;
 using Proteomics.ProteolyticDigestion;
 using UsefulProteomicsDatabases;
 using Stopwatch = System.Diagnostics.Stopwatch;
+using UsefulProteomicsDatabases.DecoyGeneration;
 
 namespace Test.DatabaseTests
 {
