+++ mzLib/Test/TestDigestionMotif.cs
@@ -15,6 +15,7 @@ using Omics.Fragmentation;
 using Omics.Modifications;
 using UsefulProteomicsDatabases;
 using Stopwatch = System.Diagnostics.Stopwatch;
+using UsefulProteomicsDatabases.DecoyGeneration;
 
 namespace Test
 {
