+++ mzLib/Test/TestPeptideWithSetMods.cs
@@ -17,6 +17,7 @@ using Omics.Modifications;
 using Transcriptomics.Digestion;
 using UsefulProteomicsDatabases;
 using Stopwatch = System.Diagnostics.Stopwatch;
+using UsefulProteomicsDatabases.DecoyGeneration;
 
 namespace Test
 {
