+++ mzLib/Test/TestProteinDatabase.cs
@@ -7,6 +7,7 @@ using System.IO;
 using System.Linq;
 using Omics.Modifications;
 using UsefulProteomicsDatabases;
+using UsefulProteomicsDatabases.DecoyGeneration;
 
 namespace Test
 {
