+++ mzLib/Test/Transcriptomics/TestDbLoader.cs
@@ -11,6 +11,7 @@ using UsefulProteomicsDatabases.Transcriptomics;
 using UsefulProteomicsDatabases;
 using Transcriptomics;
 using Omics;
+using UsefulProteomicsDatabases.DecoyGeneration;
 
 namespace Test.Transcriptomics
 {
